requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
python-json-logger==2.0.7
azure-identity==1.15.0
openai==1.12.0
sqlalchemy==2.0.25
pyodbc==5.0.1
